Keep OS CLI itx commands as functions
